@@ -19,6 +19,7 @@ import '../models/selection.dart';
 import '../service/registry.dart';
 
 class DeleteArtifactForm extends StatefulWidget {
+  const DeleteArtifactForm({super.key});
   @override
   DeleteArtifactFormState createState() => DeleteArtifactFormState();
 }
@@ -79,8 +80,8 @@ class DeleteArtifactFormState extends State<DeleteArtifactForm> {
   @override
   Widget build(BuildContext context) {
     if (artifactManager?.value == null) {
-      print("building while empty");
-      return Card();
+      debugPrint("building while empty");
+      return const Card();
     } else {
       // Build a Form widget using the _formKey created above.
       final artifact = artifactManager!.value!;
@@ -91,13 +92,13 @@ class DeleteArtifactFormState extends State<DeleteArtifactForm> {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: <Widget>[
-            Text("Delete this artifact?"),
+            const Text("Delete this artifact?"),
             Text(artifact.name),
             Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: ElevatedButton(
-                  child: Text("No, Cancel"),
+                  child: const Text("No, Cancel"),
                   onPressed: () {
                     Navigator.of(context, rootNavigator: true).pop();
                   },
@@ -106,7 +107,7 @@ class DeleteArtifactFormState extends State<DeleteArtifactForm> {
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: ElevatedButton(
-                  child: Text("Yes, Delete it"),
+                  child: const Text("Yes, Delete it"),
                   onPressed: () {
                     delete(context);
                     Navigator.of(context, rootNavigator: true).pop();
@@ -124,7 +125,7 @@ class DeleteArtifactFormState extends State<DeleteArtifactForm> {
     Selection? selection = SelectionProvider.of(context);
     if (artifactManager?.value != null && _formKey.currentState!.validate()) {
       final artifact = artifactManager!.value!.deepCopy();
-      print("deleting $artifact");
+      debugPrint("deleting $artifact");
       String subject = artifact.subject;
       artifactManager?.delete(artifact.name).then((x) {
         selection!.notifySubscribersOf(subject);
