@@ -22,6 +22,7 @@ import '../service/registry.dart';
 import '../helpers/errors.dart';
 
 class EditArtifactForm extends StatefulWidget {
+  const EditArtifactForm({super.key});
   @override
   EditArtifactFormState createState() => EditArtifactFormState();
 }
@@ -84,7 +85,7 @@ class EditArtifactFormState extends State<EditArtifactForm> {
   @override
   Widget build(BuildContext context) {
     if (artifactManager?.value == null) {
-      return Card();
+      return const Card();
     } else {
       // Build a Form widget using the _formKey created above.
       final artifact = artifactManager!.value!;
@@ -109,13 +110,13 @@ class EditArtifactFormState extends State<EditArtifactForm> {
                   return null;
                 },
               ),
-              subtitle: Text("value (string)"),
+              subtitle: const Text("value (string)"),
             ),
             Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: ElevatedButton(
-                  child: Text("Cancel"),
+                  child: const Text("Cancel"),
                   onPressed: () {
                     Navigator.of(context, rootNavigator: true).pop();
                   },
@@ -124,7 +125,7 @@ class EditArtifactFormState extends State<EditArtifactForm> {
               Padding(
                 padding: const EdgeInsets.all(8.0),
                 child: ElevatedButton(
-                  child: Text("Save"),
+                  child: const Text("Save"),
                   onPressed: () {
                     save(context);
                     Navigator.of(context, rootNavigator: true).pop();
