@@ -22,31 +22,30 @@ import '../components/split_view.dart';
 import '../helpers/media.dart';
 
 class Home extends StatelessWidget {
+  const Home({super.key});
   @override
   Widget build(BuildContext context) {
     final Selection selection = Selection();
     return SelectionProvider(
       selection: selection,
-      child: Container(
-        child: Column(
-          children: [
-            Expanded(
-              child: narrow(context)
-                  ? ProjectListCard(
-                      singleColumn: true,
-                    )
-                  : CustomSplitView(
-                      viewMode: SplitViewMode.Horizontal,
-                      initialWeight: 0.33,
-                      view1: ProjectListCard(
-                        singleColumn: false,
-                      ),
-                      view2: ProjectDetailCard(selflink: true, editable: true),
+      child: Column(
+        children: [
+          Expanded(
+            child: narrow(context)
+                ? const ProjectListCard(
+                    singleColumn: true,
+                  )
+                : const CustomSplitView(
+                    viewMode: SplitViewMode.Horizontal,
+                    initialWeight: 0.33,
+                    view1: ProjectListCard(
+                      singleColumn: false,
                     ),
-            ),
-            BottomBar(),
-          ],
-        ),
+                    view2: ProjectDetailCard(selflink: true, editable: true),
+                  ),
+          ),
+          const BottomBar(),
+        ],
       ),
     );
   }
