@@ -16,14 +16,14 @@ import 'package:registry/registry.dart';
 
 extension Display on ApiSpec {
   String nameForDisplay() {
-    return this.name.split("/").last;
+    return name.split("/").last;
   }
 
   String routeNameForDetail() {
-    return "/" + this.name.replaceAll("/locations/global", "");
+    return "/${name.replaceAll("/locations/global", "")}";
   }
 
   String routeNameForArtifacts() {
-    return "/" + this.name.replaceAll("/locations/global", "") + "/artifacts";
+    return "/${name.replaceAll("/locations/global", "")}/artifacts";
   }
 }
