@@ -26,7 +26,7 @@ import '../helpers/title.dart';
 
 class DeploymentDetailPage extends StatelessWidget {
   final String? name;
-  DeploymentDetailPage({this.name});
+  const DeploymentDetailPage({super.key, this.name});
 
   @override
   Widget build(BuildContext context) {
@@ -44,17 +44,17 @@ class DeploymentDetailPage extends StatelessWidget {
       child: DefaultTabController(
         length: 2,
         initialIndex: 1,
-        animationDuration: Duration(milliseconds: 100),
+        animationDuration: const Duration(milliseconds: 100),
         child: Scaffold(
           appBar: AppBar(
             centerTitle: true,
             title: Text(
-              pageTitle(this.name) ?? "Deployment Details",
+              pageTitle(name) ?? "Deployment Details",
             ),
             actions: <Widget>[
               homeButton(context),
             ],
-            bottom: TabBar(
+            bottom: const TabBar(
               tabs: [
                 Tab(text: "Details"),
                 Tab(text: "Artifacts"),
@@ -66,13 +66,13 @@ class DeploymentDetailPage extends StatelessWidget {
               Expanded(
                 child: TabBarView(
                   children: [
-                    DeploymentDetailCard(editable: true),
+                    const DeploymentDetailCard(editable: true),
                     narrow(context)
-                        ? ArtifactListCard(
+                        ? const ArtifactListCard(
                             SelectionProvider.deployment,
                             singleColumn: true,
                           )
-                        : CustomSplitView(
+                        : const CustomSplitView(
                             viewMode: SplitViewMode.Horizontal,
                             view1: ArtifactListCard(
                                 SelectionProvider.deployment,
@@ -85,7 +85,7 @@ class DeploymentDetailPage extends StatelessWidget {
                   ],
                 ),
               ),
-              BottomBar(),
+              const BottomBar(),
             ],
           ),
         ),
