@@ -17,14 +17,15 @@ import '../application.dart';
 import '../components/home.dart';
 
 class HomePage extends StatelessWidget {
+  const HomePage({super.key});
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         centerTitle: true,
-        title: Text(applicationName),
+        title: const Text(applicationName),
       ),
-      body: Home(),
+      body: const Home(),
     );
   }
 }
