@@ -29,7 +29,7 @@ import '../helpers/title.dart';
 
 class ProjectDetailPage extends StatelessWidget {
   final String? name;
-  ProjectDetailPage({this.name});
+  const ProjectDetailPage({super.key, this.name});
 
   @override
   Widget build(BuildContext context) {
@@ -44,17 +44,17 @@ class ProjectDetailPage extends StatelessWidget {
       child: DefaultTabController(
         length: 3,
         initialIndex: 1,
-        animationDuration: Duration(milliseconds: 100),
+        animationDuration: const Duration(milliseconds: 100),
         child: Scaffold(
           appBar: AppBar(
             centerTitle: true,
             title: Text(
-              pageTitle(this.name) ?? "Project Details",
+              pageTitle(name) ?? "Project Details",
             ),
             actions: <Widget>[
               homeButton(context),
             ],
-            bottom: TabBar(
+            bottom: const TabBar(
               tabs: [
                 Tab(text: "Details"),
                 Tab(text: "APIs"),
@@ -69,8 +69,8 @@ class ProjectDetailPage extends StatelessWidget {
                   children: [
                     ProjectDetailCard(editable: (root() == "/")),
                     narrow(context)
-                        ? ApiListCard(singleColumn: true)
-                        : CustomSplitView(
+                        ? const ApiListCard(singleColumn: true)
+                        : const CustomSplitView(
                             viewMode: SplitViewMode.Horizontal,
                             initialWeight: 0.33,
                             view1: ApiListCard(singleColumn: false),
@@ -80,11 +80,11 @@ class ProjectDetailPage extends StatelessWidget {
                             ),
                           ),
                     narrow(context)
-                        ? ArtifactListCard(
+                        ? const ArtifactListCard(
                             SelectionProvider.project,
                             singleColumn: true,
                           )
-                        : CustomSplitView(
+                        : const CustomSplitView(
                             viewMode: SplitViewMode.Horizontal,
                             initialWeight: 0.5,
                             view1: ArtifactListCard(
@@ -99,7 +99,7 @@ class ProjectDetailPage extends StatelessWidget {
                   ],
                 ),
               ),
-              BottomBar(),
+              const BottomBar(),
             ],
           ),
         ),
